pyproject.toml
[project]
name = "whisperlm"
version = "1.0.0"
description = "智能语音转文字服务：WhisperX 精确转录 + LLM 语义优化"
readme = "README.md"
requires-python = ">=3.10,<3.12"
license = { text = "MIT" }
dependencies = [
    "aiofiles>=25.1.0",
    "fastapi>=0.128.0",
    "hf-transfer>=0.1.9",
    "httpx>=0.28.1",
    "loguru>=0.7.2",
    "openai>=2.15.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "pyyaml>=6.0.3",
    "torch>=2.1.0,<2.5.0",
    "torchaudio>=2.1.0,<2.5.0",
    "uvicorn>=0.40.0",
    "whisperx",
    "demucs>=4.0.0",
    "soundfile>=0.12.0",
]

[project.scripts]
whisperlm = "whisperlm.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.sdist]
include = ["src/whisperlm"]

[tool.hatch.build.targets.wheel]
packages = ["src/whisperlm"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.uv]
override-dependencies = [
    "torch>=2.1.0,<2.5.0",
    "torchaudio>=2.1.0,<2.5.0",
    "triton>=2.1.0,<3.0.0",
]

[tool.uv.sources]
whisperx = { git = "https://github.com/m-bain/whisperX.git" }
